+++ Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/Public/Core/AccelByteServerSettings.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2019 - 2020 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
@@ -23,6 +23,7 @@ public:
 	FString IamServerUrl;
 	FString DSMServerUrl;
 	FString StatisticServerUrl;
+	FString PlatformServerUrl;
 };
 
 } // Namespace AccelByte
@@ -63,6 +64,9 @@ public:
 
 	UPROPERTY(EditAnywhere, GlobalConfig, Category = "AccelByte Server | Settings")
 	FString StatisticServerUrl;
+
+	UPROPERTY(EditAnywhere, GlobalConfig, Category = "AccelByte Server | Settings")
+	FString PlatformServerUrl;
 };
 
 
@@ -96,6 +100,9 @@ public:
 	static FString GetStatisticServerUrl();
 
 	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
+	static FString GetPlatformServerUrl();
+
+	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
 	static void SetClientId(const FString& ClientId);
 
 	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
@@ -109,11 +116,13 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
 	static void SetIamServerUrl(const FString& IamServerUrl);
-	
+
 	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
 	static void SetDSMServerUrl(const FString& DSMServerUrl);
 
 	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
 	static void SetStatisticServerUrl(const FString& StatisticServerUrl);
-};
 
+	UFUNCTION(BlueprintCallable, Category = "AccelByte Server | Settings")
+	static void SetPlatformServerUrl(const FString& PlatformServerUrl);
+};