+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.6.1 (2020-02-07)
+## Fixed
+- Game Client SDK
+  - Fix LoginWithOtherPlatform PlatformType
+
 ## 2.6.0 (2020-01-29)
 ## Added
 - Game Client SDK